refactor: structural refactoring
